Change the way errors are handled.
Add the UnsuccessfulResponseBody model and UnsuccessfulHttpResponseException, which serves as a container of information about errors that occur when requests are sent.

Add UnsuccessfulHttpResponseExceptionFactory, which allows you to create an exception object based on the http response from the server. Refactor SbermarketApiClientV3 so that it uses UnsuccessfulHttpResponseExceptionFactory when handling exceptions.

Write comments to some classes.